requirements.txt
Flask==2.0.2
PyJWT==2.1.0
Werkzeug==2.0.2
